(#9297) - move ConfiguredResourceAttributesHolder to javaagent-extension-api
