+++ instrumentation/log4j/log4j-context-data/log4j-context-data-2.17/library-autoconfigure/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 base.archivesName.set("${base.archivesName.get()}-autoconfigure")
 
 dependencies {
-  compileOnly(project(":javaagent-bootstrap"))
+  compileOnly(project(":javaagent-extension-api"))
   library("org.apache.logging.log4j:log4j-core:2.17.0")
 
   testImplementation(project(":instrumentation:log4j:log4j-context-data:log4j-context-data-common:testing"))
